more misc updates
